Improving code  - Creating a test file.  - Recfactoring curry code.
